refactor StatusUiState

+++ commonbiz/status-ui/src/commonMain/kotlin/com/zhangke/fread/status/ui/label/StatusBottomInteractionLabel.kt
@@ -27,8 +27,8 @@ import org.jetbrains.compose.resources.stringResource
 @Composable
 fun StatusBottomInteractionLabel(
     modifier: Modifier,
-    boostedCount: Int,
-    favouritedCount: Int,
+    boostedCount: Long,
+    favouritedCount: Long,
     style: StatusStyle,
     onBoostedClick: () -> Unit,
     onFavouritedClick: () -> Unit,
